oui-httpd: separates lua library
