+++ applications/oui-app-cron/files/rpc/crontab.lua
@@ -1,3 +1,5 @@
+local utils = require "oui.utils"
+
 local M = {}
 
 function M.get()
