+++ applications/oui-app-upgrade/files/rpc/upgrade.lua
@@ -1,3 +1,5 @@
+local utils = require "oui.utils"
+
 local M = {}
 
 function M.verify(params)
