+++ oui-httpd/Makefile
@@ -39,12 +39,13 @@ endef
 
 define Package/oui-httpd/install
 	$(INSTALL_DIR) $(1)/usr/sbin $(1)/etc/init.d $(1)/etc/config \
-		$(1)/usr/lib/oui-httpd/rpc $(1)/usr/lib/oui-httpd/lib \
-		$(1)/etc/oui-httpd 
+		$(1)/usr/lib/oui-httpd/rpc $(1)/etc/oui-httpd $(1)/usr/lib/lua/oui/utils
 	$(INSTALL_CONF) ./files/oui-httpd.config $(1)/etc/config/oui-httpd
 	$(INSTALL_BIN) ./files/oui-httpd.init $(1)/etc/init.d/oui-httpd
 	$(INSTALL_BIN) $(PKG_BUILD_DIR)/oui-httpd $(1)/usr/sbin
-	$(INSTALL_DATA) ./files/utils.lua $(1)/usr/lib/oui-httpd/lib
+	$(CP) $(PKG_BUILD_DIR)/json.so $(1)/usr/lib/lua/oui
+	$(CP) $(PKG_BUILD_DIR)/utils.so $(1)/usr/lib/lua/oui/utils/core.so
+	$(INSTALL_DATA) ./files/utils.lua $(1)/usr/lib/lua/oui/
 	for f in `ls ./files/rpc`; do \
 		$(INSTALL_DATA) ./files/rpc/$$$$f $(1)/usr/lib/oui-httpd/rpc/$$$${f%.*}; \
 	done
