+++ oui-httpd/files/rpc/oui.lua
@@ -1,5 +1,7 @@
 local uci = require "uci"
 local sqlite3 = require "lsqlite3"
+local utils = require "oui.utils"
+local cjson = require "oui.json"
 
 local M = {}
