+++ oui-httpd/files/rpc/system.lua
@@ -1,3 +1,5 @@
+local utils = require "oui.utils"
+
 local M = {}
 
 function M.diskfree()
